requirements.txt
numpy
scipy
scikit-image
hickle
Pillow
pandas